Refactor: remove stringify utils as public APIs
