+++ src/utils/index.ts
@@ -5,4 +5,3 @@ export { convertTransactionStructToCommitmentSummary } from './commitment';
 export * from './ecies';
 export * from './encryption/aes';
 export { getPublicViewingKey, verifyED25519 } from './keys-utils';
-export * from './stringify';
